Use Scan Policy configured in Policy tab in Active Scan dialogue
Change class CustomScanDialog to obtain the scan policy configured in
the Policy tab when "advanced options" is enabled. Move Policy tab
related code to a nested class as it makes it easier to use (its also
responsible to update the category panels with the selected scan policy,
ensuring that the changes done in those panels are reflected in the
selected scan policy).
Change class PolicyAllCategoryPanel to allow interested parties to be
notified of changes in the selected scan policy and to reset/reload the
policies and select a scan policy (requirements per above changes).
Issue introduced/revealed with the changes done in #2148 (for issue
#2112), since the scan policies no longer share the Plugin instances
(therefore the scan policy being obtained in the save method would not
have the changes done in the Policy tab).