Allow to active scan the HTTP headers of all requests
Change classes ScannerParam and OptionsVariantPanel to allow to specify
if the HTTP headers of all requests should be scanned, programmatically
and through the GUI, respectively.
Change class VariantHeader to use the new option from ScannerParam
allowing to scan (or not) the HTTP headers of all requests.
Fix #1959 - Allow to active scan headers of all requests