Initialise Technology tab with selected context in Active Scan dialogue
Change CustomScanDialog to use TechnologyTreePanel (an extracted class
from ContextTechnologyPanel) which reduces code duplication, fixes
selection issues (for example, #1708 and #2110) and correctly initialise
the tree with the technologies of the selected context. The technologies
of the context were not being selected as it was iterating over the
wrong variable ("iter" instead of "iter2"), also the technologies were
not being deselected (when excluded).