Install new versions of the add-ons after downloading with -addonupdate
Change ExtensionAutoUpdate to call the method installNewExtensions()
after downloading the new versions of the add-ons, so that the add-ons
are effectively updated.
Extract a method that waits for the downloads to finish and install
them, as that's done by all ExtensionAutoUpdate command line arguments.
Log as INFO when an add-on is being installed (so that by default that
information is logged).