Disallow Spider scans when ZAP is in Safe (or Protected) mode
Change method ExtensionSpider.sessionModeChanged(Mode) to disable the
"Spider.." menu when the mode is changed to Safe (and enable otherwise),
it would allow to start the scans even in "Safe" mode. Also, change to
not start the scans if the mode is "Safe" or if the target is not in
scope when in "Protected" mode.
Change SpiderDialog to inform and do not start the scan when in Safe or
Protected mode, if the target is not in scope.