Change Spider to log the seeds with DEBUG level and keep them unique
Change Spider class to use a LinkedHashSet to keep the seeds, to avoid
repetitions while keeping the same iteration order as the ArrayList. The
seeds are repeated when they are added by the built-in parsers, per seed
added, for example:
 Adding seed for spider: http://example.com/
(following seeds added by built-in parsers)
 Adding seed for spider: http://example.com/robots.txt
 Adding seed for spider: http://example.com/sitemap.xml
 Adding seed for spider: http://example.com/.svn/entries
 Adding seed for spider: http://example.com/.svn/wc.db
 Adding seed for spider: http://example.com/.git/index

Change the log level, of the seeds added, from INFO to DEBUG as it logs
too much messages when spidering a target more than once (since
previously visited pages are added as seed, plus seeds added by built-in
parsers).