Change Git Spider parser to DEBUG normal case and WARN instead of ERROR
Change class SpiderGitParser to not throw an exception for normal case
and log with DEBUG level (instead of ERROR). For example, when the
response is not a Git index file, it would throw an exception and log as
error:
 ERROR SpiderParser  - An error occurred trying to parse Git url
 'http://example.com/.git/index': java.lang.Exception: The file
 '/.git/index' could not be parsed as a Git Index file due to unexpected
 content
although this is a normal case that happens frequently (when the seeds
are automatically added for the parser which might result in a 404 type
response).
For issues during the parsing of the index log exceptions with WARN
level, as well when the file does not have the expected path.