Allow to enable/disable database's recovery log (HSQLDB)
Add option to "Database" options panel to allow to enable/disable
database's recovery log, by changing the class OptionsDatabasePanel.
Change class DatabaseParam to load/persist the new option from the
configuration file.
Change class ParosDatabase to allow to set the database options so that
they can be checked and applied (in this case the option recovery log)
by ParosDatabaseServer when connecting to new or existing databases.
Change class ParosDatabaseServer to enable/disable recovery log
depending on the state of the new option.
Change method Model.init(ControlOverrides) to load the configuration
file before creating the databases so that the databases can use the
configurations and also change it to set the database configurations to
ParosDatabase instance (HSQLDB implementation).
Fix #1958 - Allow to disable database (HSQLDB) log