Change "sessions" API view to not use tokens' names as XML element names
The "sessions" API view will now return as tokens of a session:
 <tokens type="set">
   <token type="set">
     <name>SESSID</name>
     <value>123</value>
     <domain>example.com</domain>
     <path>/</path>
     <secure>false</secure>
   </token>
 </tokens>
instead of:
 <tokens type="set">
   <SESSID>SESSID=123</SESSID>
 <tokens>
preventing exceptions when the name of the session tokens are not valid
XML element names and normalising its structure (no longer dependent on
tokens' names).
Change class HttpSessionsAPI to use a custom ApiResponseSet, to produce
the newer XML structure (class TokenValuesResponseSet), also extract a
method that creates a response with session's data, to reduce code
duplication.
Fix #2246 - Error in sessions view with tokens