Change "params" API view to filter the results with the given site
Change class ParamsAPI to use the optional "site" parameter to filter
the results returned in the API response. Extract a method that creates
the API response to reduce code duplication, it's used for "all results"
and the filtered ones.
Change class ExtensionParams to allow to check if a site already exists
and if it has parameters, to allow the API to return an error if it
doesn't and to not create the (empty) response if it does not have any
parameters (also, change class SiteParameters to allow to check if it
has or not parameters).
Move the method HttpSessionsApi.getAuthority(String) to class ApiUtils,
it's now used by several classes (HttpSessionsApi and ParamsAPI) and
change it to return default port for HTTP scheme to not require the
callers to check and add the port. Also, move the tests to a new class,
ApitUtilsUnitTest, and update them to conform to new behaviour.