Fix NullPointerException in VariantCookie implementation
Change class VariantCookie to conform to the expected behaviour when the
parameter's (that is, the cookie) name and value are set to null, which
indicates the parameter should be removed, preventing
NullPointerExceptions, for example:
ERROR org.zaproxy.zap.extension.ascanrules.TestParameterTamper -
 Error occurred while scanning a message:
[...]
java.lang.NullPointerException
 at java.net.HttpCookie.<init>(HttpCookie.java:143)
 at java.net.HttpCookie.<init>(HttpCookie.java:139)
 at VariantCookie.setParameter(VariantCookie.java:125)
 at VariantCookie.setParameter(VariantCookie.java:92)
 at AbstractAppParamPlugin.setParameter(AbstractAppParamPlugin.java:259)
 at TestParameterTamper.scan(TestParameterTamper.java:129)
[...]

Also, change the class to:
 - Manually parse and create the cookie header to correctly handle null
 names and values (which would lead to similar exceptions if used with
 HttpCookie class);
 - Remove (now) redundant null checks from methods getEscapedValue and
 getUnescapedValue;
 - Update/add JavaDoc to class and methods.

Change class NameValuePair to add the methods hashCode(), equals(Object)
for proper use in collections and comparisons, and method toString() for
help with debug and remove redundant instance variable initialisations.
Add tests to assert the expected behaviour of classes VariantCookie and
NameValuePair.