spider: use all information/data available when seeding
Change SpiderThread to use the start URI and start node as well as the
context nodes or all nodes in scope.

The spider will be seeded with following constraints:
 - If a context is provided:
  - Start URI, if in context;
  - Start node, if in context;
  - All nodes in the context;
 - If spidering just in scope:
  - Start URI, if in scope;
  - Start node, if in scope;
  - All nodes in scope;
 - If there's no context/scope restriction:
  - Start URI;
  - Start node, also:
    - Child nodes, if spidering "recursively".

Differences to previous seeding behaviour:
 - The start URI and start node are always used (if constraints met);
 - The nodes of the provided context are always used (previously, the
 nodes might not be used if a start node was also provided);

Seeding with context and "just in scope" are mutually exclusive, taking
precedence the former (same behaviour as before this change).