spider: use the broadest target component for scan's display name
Change spider to create the display name for the scans using all the
components of the target and custom configurations, with the following
order:
 - Context;
 - Just in scope;
 - Start node;
 - Start URL.

So, for example, if a start node is specified along with a context it
will display the context name instead of the start node (which is more
specific, narrow target when spidering).

Change class ExtensionSpider to follow the above rules using a new
method to start the scans, to ensure that all scans use the same display
name rules (the old method that allows to define the display name is
still available but its use is discouraged in favour of using the new
method).
Change classes ExtensionSpider, SpiderAPI and SpiderDialog to use the
new method, for the latter class also remove the creation of the custom
display name.