Fix exceptions related to Http Sessions extension/tab
Change method HttpSessionsSite.getMatchingHttpSession(...) to create a
copy of the sessions (in a synchronised block) and pass the copy to a
method of CookieBasedSessionManagementHelper to prevent the exception
ConcurrentModificationException (which happened while iterating/copying
the original sessions collection in the method of the latter class).
Change class HttpSessionsTableModel to do the modifications to sessions
collection and notifications to the GUI in the EDT to prevent leaving
the view in inconsistent state, which would lead to exceptions like
"IndexOutOfBoundsException: Invalid range", when sorting the entries
added to the "Http Sessions" table.
Change method ExtensionHttpSessions.onHttpResponseReceive(...) to not
process the responses of fuzzed messages, which can potentially generate
thousands of "synthetic" sessions in ZAP (like the active scanner,
already excluded).
Fix #2142 - Fuzzer throwing exceptions