spider: preserve query parameters with same name when canonicalising URL
Change class URLCanonicalizer to keep query parameters with same name
to preserve (or attempt to) the meaning of the original URL.
For example, with old behaviour URLs like:
 http://example.com/?array[]=1&array[]=2
 http://example.com/?IDs=1&IDs=2
would loose one of the query parameters, which could result in different
content being served.
Add tests to assert the expected behaviour.