Change "context" view to wrap the context data with an object
Change class ContextAPI to wrap the context data with an object (as all
views do), called "context", instead of returning the context data
directly, so that the Python ZAP API client is able to correctly read
the data, which expects (and removes) the wrapper object.

The data returned is now, for example:
 {"context":{"id":"1", ..., "inScope":"true","loggedOutPattern":""}}

instead of:
 {"id":"1", ..., "inScope":"true","loggedOutPattern":""}

This change breaks existing custom clients that consume the JSON format.
The structure of XML format is unchanged.

Fix #2115 - Python API context.context("MyContext") is broken