Fix NPE in JCheckBoxTree's renderer
Change class CheckBoxCellRenderer to not attempt to convert the user
object, of the tree node being rendered, to a string if there's none
(that is, null). Likely to happen in a tree with a non-visible root,
which might still be rendered for accessibility reasons.
Fix #2314 - Cannot add several payloads to fuzzer (NullPointerException)