Include scanners' quality in API progress view
Change class ActiveScanAPI to include the quality of the scanners in the
"scanProgress" view.
The response will now include a field called "quality" for each scanner,
with the corresponding quality ("alpha", "beta" and "release").

Other code change, extracted a method to reduce code duplication when
creating a scanner progress entry, for the "scanProgress" response.

Fix #2183 - API: Include quality in the ascan/progress response