Allow to spider a site's subtree
Add a new option to Spider dialogue ("Spider Subtree Only") and a
parameter to spider ZAP API ("subtreeOnly") to allow to spider just a
site's subtree. In both cases it's required a URI (manually set or
previously accessed), that's used as the subtree to spider. It's only
used the scheme, host, port and path component of the URI when
considering if a resource found while spidering is under or not the
specified subtree. For example, if the seed URI is:
 http://example.com:8080/path/?a=b#c
it's used only:
 http://example.com:8080/path/

The spider will exclude any resources found not under the subtree by
using a new FetchFilter, HttpPrefixFetchFilter, that allows to filter
URIs based on a URI prefix, in this use case the subtree URI.
Add tests to assert the expected behaviour of HttpPrefixFetchFilter.
Fix #2274 - Allow to spider just a site's subtree