Add Outbound Proxy Password Peak
Add a checkbox to allow the user to show their password.
Password is currently shown as bullets/pips by default
to prevent shoulder surfing. The checkbox setting is 
temporary and always defaults to unselected/disabled
on each display of the dialog(s). Changes applied to 
both the connection options panel and prompt at 
startup dialog.