Do not require a restart to show/hide the tool bar
Change class MainFrame to always add the main tool bar, so that it can
be shown/hidden at any time, also add a method to allow to change its
visible state.
Change class View to apply the main tool bar visibility option when it
is initialised and add a method to allow to change the visible state of
the main tool bar (which calls the method of MainFrame class).
Change class MenuToolsControl to update the View with the main tool bar
visibility option, after the options are set.
Change class BreakPanel to take into account the changes done to
visibility of the main tool bar (to show/hide the buttons in the break
panel if the main tool bar is hidden/shown and the option is to show or
not the buttons in the tool bar).
Change class OptionsParamView to use boolean (for external code) instead
of an int for the visibility option of the main tool bar and deprecate
the methods that use int.
Change class OptionsViewPanel to use the boolean methods of class
OptionsParamView.

Fix #2368 - Do not require a restart to show/hide the main tool bar