Change API authorization view to wrap its object
Change class AuthorizationAPI to wrap the authorization data with an
object (as all views do), called "authorizationDetectionMethod", instead
of returning the authorization data directly, so that the Python ZAP API
client is able to correctly read the data, which expects (and removes)
the wrapper object.

The data returned is now, for example:
 {"authorizationDetectionMethod":{"statusCode":"-1"..."headerRegex":""}}

instead of:
 {"statusCode":"-1"..."headerRegex":""}

This change breaks existing custom clients that consume the JSON format.
The structure of XML format is unchanged.