Show HTTP messages in the Spider tab
Add a button to tool bar of Spider tab to allow to see the HTTP messages
sent while spidering. When the button is selected the panel shows two
tabs, one for the URLs and the other for the HTTP messages, otherwise it
just shows the table with the URLs found, as before.