Disable "Show in History Tab" if message not valid
Disable the menu item "Show in History Tab" if the type of the selected
message is not valid for the History tab (only manual or proxied
requests are shown in that tab). Prevents the menu item from being
enabled for other message types (for example, scanner, spider) which
would have no effect other than showing/focusing the History tab.

Change class (extension.stdmenus.)PopupMenuShowInHistory to not be
enabled for the mentioned types and to not be shown in the Spider panel,
now that it shows HTTP messages.
Deprecate class (extension.history.)PopupMenuShowInHistory, replaced
with the former class in a previous commit.