Remove unnecessary log warnings
Change method ExtensionGlobalExcludeURL.optionsLoaded() to remove the
existing log warning, not needed.
Change method OptionsGlobalExcludeURLPanel.saveParam(Object) to remove
the log warning (by moving the logged message, the ignored regular
expressions, to existing debug log statement and wrap it with a
"isDebugEnabled"), as with the previous warning it's not needed (as
warning).