Reverse the uninstallation of add-on's extensions
Change method AddOnInstaller.uninstallAddOnExtensions(...) to reverse
the order of uninstallation of the extensions of the add-ons, extensions
that are declared later in the manifest (ZapAddOn.xml) might depend on
extensions declared earlier, which might lead to exceptions if those are
uninstalled first. This is similar to the uninstallation order of other
components of an add-on, for example, active and passive scanners are
installed after extensions, when uninstalling the add-on they are
unsintalled before the extensions.

Fix #2436 - Unable to dynamically uninstall WebSockets add-on