Raise SiteMap Add or Remove Events
SiteMap 
- Add private method handleEvent: Called from findAndAddChild,
findAndAddLeaf, removeNodeFromParent. 
- Add private convenience method publishEvent that accepts an event and
node to publish the necessary event (leveraging
SiteMapEventPublisher). Node events are always published,
site events are also published when the parent is the root of the tree.
- New overridden removeNodeFromParent method that extracts the
node's parent, calls the super method, and makes the call for the event
to be handled.

SiteMapEventPublisher
- Added site events (node events already existed).

Fixes zaproxy/zaproxy#1171