Fix exception while opening Active Scan dialogue
Change method OptionsParamView.setLocale(Locale) to save the locale as
String instead of Locale (call existing method setLocale(String),
preventing the issue if the OptionsParamView is (re)loaded.
Remove instance variable CustomScanDialog.optionsParam and corresponding
load of configurations (which was triggering the issue) when the
dialogue is reset, the optionsParam is not being used. Also, change to
use the method ScannerParam.clone() instead of doing the clone manually.

Fix #2440 - Exception while opening the Active Scan dialogue