Allow to copy active scan's progress details
Change class ScanProgressDialog to allow to copy (to the clipboard) the
contents of the table that shows the progress of the active scan. The
change adds a button (to the footer of the dialogue, same place as the
close button) that does the actual copy of contents to the clipboard.
Change classes ScanProgressActionIcon and ScanProgressItem to implement
the method toString(), used when copying the data.

Fix #1864 - Copy scan progress as text to clipboard