Adjust ProxyThread log levels and log read timeout
Change class ProxyThread to adjust the log level of exceptions caused by
listeners (a probable issue/bug) from WARN to ERROR and change some log
WARN levels to DEBUG (ones that are not an issue/bug). Also, change to
log the timeout with WARN level, per discussion in #2438.
Change log4j.properties file to remove the override that changed the
default log level (of the class ProxyThread) to ERROR.