Allow to access a URL through the ZAP API
Add a convenient and simple action to CoreAPI, called "accessUrl", that
allows to access a URL, optionally following redirections. Since the URL
is accessed from within ZAP it does not require external applications to
trust ZAP's root CA certificate when accessing HTTPS URLs. As with the
other actions that allow to send requests this one also returns the
request sent and response received (and redirections followed, if any).