Always use charset set when changing the HTTP body
Change class HttpBody to always use the charset set when creating the
String representation of the body (it could happen in some cases that it
would use the default charset instead of the one set, leading to wrong
encoding of the characters). Also, fix undocumented exceptions/behaviour
and change to use StandardCharsets/Charset (instead of String) when
obtaining the bytes from the String or converting the bytes to String.
Change class HttpResponseBody to also use StandardCharsets/Charset and
use the new code to create the String representation.
Add tests to assert the expected behaviour of classes HttpBody and
HttpResponseBody.

Fix #2487 - Wrong charset used in HTTP body