Proxy WebSocket traffic through outgoing proxy set
Change class HttpSender to also use the custom connection manager (one
that does not close the connection and exposes its socket) when sending
requests that have a connection upgrade header (e.g. WebSocket) through
an outgoing proxy.
Change HttpMethodDirector to establish a tunnel if the request has a
connection upgrade header.
Fork HttpConnection to allow to establish a tunnel when requested by
calling code (in this case when the request has a connection upgrade).

Fix #2463 - Websocket not proxied when outgoing proxy is set