Normalise Alert.equals/hashCode/compareTo
Normalise the implementation of Alert's methods equals and compareTo and
implement the method hashCode for consistency when used in different
collections.
Two alerts are considered the same if the following properties are
equal:
 - Plugin ID;
 - Name;
 - Risk;
 - Confidence;
 - URI;
 - Parameter;
 - Evidence;
 - Attack;
 - Other Info.

Change class AlertTreeModel to use the method Alert.compareTo instead of
comparing manually the alerts to ensure that the alerts shown in the
Alerts tab are the same as the ones included in the report.

Fix #2132 - Zap Report Counting Bug