Reset the charset of the body with null/empty name
Change class HttpBody to reset the charset when it's set a null/empty
charset name, to ensure that it always uses the charset set from the
header (if reset it defaults to default charset or the one determined
internally by HttpBody implementations, like HttpResponseBody).
Change tests to assert the (new) expected behaviour of class HttpBody.

Related to #2493 - Always use charset set when changing the HTTP body
(Since those changes it no longer determines, always, the charset from
the response body when the cached string is re-created.)

Fix #2517 - Wrong charset used in resent response