Honour Mode and scope when following redirections
Change classes AbstractPlugin (active scanners) and Analyser (pre active
scanner task) to honour scan's scope when following redirections (which
already includes the Mode).
Change class HttpPanelSender ("Manual Request Editor" and "Resend"
dialogues) to honour the Mode when following redirections.
Change class CoreApi to honour the Mode when sending manual requests and
following redirections, either plain text or in HAR format (as with UI
the requests are not allowed if in 'Safe' mode or in 'Protected' mode if
out of scope).
Change class HttpSender to allow to validate the redirections before
following them.

Fix #2546 - ZAP access URLs which are out of scope