Remove JAR scan for authentication/session methods
Change classes ExtensionAuthentication and ExtensionSessionManagement to
not scan the loaded JARs (core and add-ons) to load authentication and
session management methods, instead directly add/load the methods
provided by core code.
The change reduces the time required to start ZAP as it does not need to
scan all the JARs to load the methods, moreover it will also stop
loading any methods provided by add-ons (if any), the add-ons should
add/remove the methods to be properly installed/updated.

Related to #2620 - Allow to add/remove authentication and session
management methods