Check older add-ons on manual installations
Change ExtensionAutoUpdate to check if the add-on being manually
installed is a newer or an older version, to correctly inform about
required changes during the installation (for example, if other add-ons
need to be uninstalled, updated or installed). Also, inform with better
message if the same add-on (version and status) is already installed.

Fix #2707 - Manual add-on installation needs more meaningful dialog
messages