Use request method to differentiate alerts
Change Alert class to use the request method to differentiate the alerts
(in equals/hashCode/compareTo) and to include the request method when
converting to XML (which is included in the XML report).
Change CoreAPI to include the request method in the details of the
alerts.
Change HTML XSL (used for HTML report) to include the request method for
each alert.

Fix #2757 - Alerts with different request method are considered the same