Show correct error for wrong regex in ContextAPI
Change ContextAPI to throw ILLEGAL_PARAMETER (instead of INTERNAL_ERROR)
when the regular expression (for context inclusion or exclusion) is
invalid, which better indicates what the problem is.
Change Context class to not throw Exception and document the exceptions
expected to be thrown (PatternSyntaxException if the regular expression
is not valid or IllegalArgumentException if it it's null).
Change ContextExcludePanel and ContextIncludePanel to no longer catch
the exception when setting the (already validated) regular expressions.