Differentiate the source of alerts
Change the Alert class to have a source "unknown", "active", "manual",
"passive" and "tool" which is set just before the alert is raised for
active, manual and passive alerts, for remaining alerts it's considered
as raised by a tool (e.g. custom scripts). Old alerts of existing
sessions will default to "unknown" since its exact source is not
immediately known.
Change Alert tab to shown the new field of the alert, "Source:".
Change CoreAPI to return the ID of the source of the alerts.
Change databases and corresponding tables to have the source.
Change HTML, XML and MD reports to include the ID of the source of the
alerts.

Fix #2592 - Differentiate the source of alerts