Attempt to determine (String) body's charset
Change HttpBody and HttpResponseBody to attempt to determine the charset
of the contents (String) being set if the charset is unknown (that is,
it was not previously set before the contents are set).
Update tests to reflect the change in the behaviour.

Related to #2487 - Wrong charset used in HTTP body
Fix #2935 - Wrong charset used in response body if no charset set