Initialise panels when added to session dialogue
Initialise the panels when added to session dialogue if it's shown, to
ensure that the panels are in a consistent state. Also, ensure the
session dialogue has a "UI shared context" when adding the panels of the
newly added context.
The change prevents exceptions (caused by the inconsistent state of the
panels) when changes are done to the contexts (e.g. via ZAP API) while
the dialogue is shown.
Change to initialise the "regular" (i.e. non context) panels only once,
when initParam(Object) is called (already done by base class).