Add Spider scans to GUI in the EDT
Change ExtensionSpider to add the spider scans to the GUI in the EDT, to
prevent inconsistencies between EDT and other threads, which could lead
to exceptions (and a freeze of GUI caused by inconsistent internal state
of UI components).
Change SpiderScan to not create the model when adding messages if the
scan was already cleared, to prevent a leak of AlertEventConsumer(s).