Do not require status/version in add-on file name
Change AddOn class to not require the status and/or the version in the
file name of the add-on. The add-on file name just needs to have the ID
and have a ZAP extension. Also, deprecate old constructor/methods that
require the file name to have the status/version and introduce new
constructor/methods where applicable.
Change BaseZapAddOnXmlData to read the status from the manifest file of
the add-on (ZapAddOn.xml).
Remove hardcoded manifest file name (ZapAddOn.xml) from JavaDoc and code
(by using the constant from AddOn).
Change AddOnCollection to iterate just ZAP add-on files and use the new
AddOn constructor.
Change ExtensionAutoUpdate to use the new constructors/methods and to
copy the file from manual add-on installations using a normalised file
name.
Add some tests to assert the expected behaviour of AddOn class.

Fix #3090 - Be more lenient on add-on's file name format