Move Spider functionality to Spider extension
Move (that is, create new class and deprecate old one) the spider
functionality to ExtensionSpider, also, all spider related functionality
is now done through classes in the spider extension package.
The functionality "Exclude from Spider" is shared by both spiders (AJAX
and traditional) therefore it belongs to core.

Preparation for #3113 - Move Spider to an add-on