Remove Session dependency on ExtensionActiveScan
Change Session to no longer call/use ExtensionActiveScan.
Change ActiveScanController to obtain the excluded URLs (session and
global) instead of having the Session to set them.
Change ExtensionActiveScan to allow to set a list of excluded URLs and
to not change running active scans, normalising the behaviour with the
normal spider.

Part of #3129 - Move Active Scanner to an add-on