Determine charset only for text responses
Change HttpBody to allow to set whether or not the charset should be
determined.
Change HttpMessage to set to determine the charset of the (String)
response only if it's a text response.
Update tests to reflect the change in the behaviour (and fix JavaDoc
of related class).

The change prevents the incorrect body length calculation for non-text
responses (e.g. image).