Always remove/unhook the extensions
Change AddOnInstaller to remove/unhook the extension instead of letting
the extension remove/unhook itself. Avoids the need of the extension
to call the (base) unload method when overriding it (which if not done
would cause subtle issues, like the extension not being removed and the
components not being unhooked).