Issue 1634 Improve Export URLs
Exporting URLs to a file now takes into consideration what is selected
in the Sites Tree.

Refactored PopupMenuExportURLs so that it uses HistoryReferences instead
of reading from the DB, and also made it more modular.

Added PopupMenuExportSelectedURLs which extends PopupMenuExportURLs,
checks the selection (incl. multi) in the site tree and exports based on
selection. (If no selection, export all).

The revised version of this functionality no longer exports the HTTP
Method. It exports only URLs in order to be more true to the UI
experience of the user.

Fixes zaproxy/zaproxy#1634

zap-core-help\src\help\zaphelp\contents\ui\tlmenu\report.html will be
updated in a separate PR after this one is accepted.