Change to Snapshot As
Change Snapshot to allow the user to select the name. A datestamped name
is suggested based on the persisted session name along with a timestamp,
however the user can set it to anything they wish.

Accompanying help PR (https://github.com/zaproxy/zap-core-help/pull/69)
will be updated after this change is accepted.