Specify listen port and wait for passive scanning
This patch adds a new option to the zap-baseline.py script that permits
to specify the listening port of zap with the -P option, thus we can
know in advance the port that we can use as a passive proxy.

Also, the patch adds a loop to wait for the time specified with the -D
option, that permits us to run external external software using the zap
proxy (i.e selenium) and obtain the analysis of the passive scanning. In
current implementation the script finish at the same time that the
spider, that makes impossible to control the time of the passive
scanning.

Signed-off-by: Antonio Ojea <aojea@midokura.com>