Correct API's hostname check
Correct API's hostname check, by swapping its statements, return true if
matched otherwise log and return false. Also, rename the method that
does the check to not have IP in the name and allow to set the API
options to ease the tests.
Add tests to assert the expected behaviour of address checks.