Fix Params Tab Load Saved Session Issue
This fixes an issue whereby upon loading a saved session the param panel
site selector dropdown menu contained only http (:80) entries which do
not necessarily align properly to the session being loaded. 

Previously the scheme was being stripped when the site tree was
enumerated, this meant that all the https entries were not properly
added to the site selector dropdown.