Correct creation of Git/SVN spider seeds
Change Spider class to create the URIs using the escaped path, as the
URI constructor used expects the contents to be already escaped,
preventing exceptions like:
   URIException: escaped absolute path not valid

which happened if the path contained escaped characters. The code that
created the seeds was refactored to reduce code duplication, also, in
case of exception while creating the URI log all data available (not
just the base URI).