Allow add-ons to use Semantic Versioning
Change AddOn class to use semantic version by default, if an add-on is
still using integer version it's converted to a semantic version.
The integer/file version is also deprecated and should be used just the
semantic version (just referred as version).
The semver element used in the ZapAddOn.xml and ZapVersions.xml files is
deprecated in favour of using only the version element.
Add/update tests to assert the new behaviour.

Fix #3101 - Allow add-ons to use Semantic Versioning