Move RedirectionValidator to a top level class
Move RedirectionValidator to a top level class (and rename it to
HttpRedirectionValidator), also introduce HttpRequestConfig, to allow
to easily configure a HTTP request.
Change HttpSender class to use HttpRequestConfig as parameter to send
the requests (instead of directly RedirectionValidator).
Update the classes that were using RedirectionValidator to use the new
classes.