Change API view "contextList" to return a list
Change class ContextAPI to return the context names in a list, not all
names in a single string.

The data returned is now, for example:
```JSON
{"contextList":["Context 1","Context 2"]}
```
and:
```XML
<contextList type="list">
    <contextName>Context 1</contextName>
    <contextName>Context 2</contextName>
</contextList>
```

instead of:
```JSON
{"contextList":"[Context 1, Context 2]"}
```
and:
```XML
<contextList>[Context 1, Context 2]</contextList>
```

This change breaks existing consumers that extract the names from the
string.

Fix #3353 - ZAP API View:contextList method returns a string instead of
JSON list