Improve error handling when resetting the options
Change OptionsDialog to catch the exceptions thrown when resetting each
panel, and inform the user in that case. Also, changed to log the
exceptions (as error), necessary to easily resolve the issue(s) that
might occur during reset.
Remove extra/unnecessary spaces from the errors messages related to
reset and validation of the options.